Reorganize to avoid bug
If we try to do anything in the CHROOT with cmd_basic_chroot after we fix the resolv.conf file, we get DNS failures.

I want to avoid using the more full-featured cmd_chroot where possible as it increases the risk of certain bugs.

So, download the config tar ball before we mess with the resolv.conf file, using the basic cmd_basic_chroot function.